requirements.txt
pytest==8.3.3
pytest-xdist==3.6.1
requests==2.32.3
Appium-Python-Client==4.0.0
allure-pytest==2.13.5
python-dotenv==1.0.1
